SI7A AGENT

+++ access_management/models/crm_lead.py
@@ -19,7 +19,7 @@ class CrmLead(models.Model):
     def get_view(self, view_id=None, view_type="form", **options):
         result = super().get_view(view_id=view_id, view_type=view_type, **options)
 
-        if self._dg_read_all_active() and view_type in {"list", "kanban", "form"}:
+        if (self._dg_read_all_active() or self._siha_agent_read_only()) and view_type in {"list", "kanban", "form"}:
             node = etree.fromstring(result["arch"])
             node.set("create", "0")
             node.set("edit", "0")
@@ -28,23 +28,28 @@ class CrmLead(models.Model):
 
         return result
 
+    def _bd_manager_no_delete(self) -> bool:
+        return (not self.env.su) and self.env.user.has_group("access_management.group_bd_manager")
+
+    def _siha_agent_read_only(self) -> bool:
+        return (not self.env.su) and self.env.user.has_group("access_management.group_siha_agent")
+
     @api.model_create_multi
     def create(self, vals_list):
-        if self._dg_read_all_active():
+        if self._dg_read_all_active() or self._siha_agent_read_only():
             raise AccessError(_("Vous n'êtes pas autorisé à créer des opportunités/pistes CRM."))
         return super().create(vals_list)
 
     def write(self, vals):
-        if self._dg_read_all_active():
+        if self._dg_read_all_active() or self._siha_agent_read_only():
             raise AccessError(_("Vous n'êtes pas autorisé à modifier des opportunités/pistes CRM."))
         return super().write(vals)
-
-    def _bd_manager_no_delete(self) -> bool:
-        return (not self.env.su) and self.env.user.has_group("access_management.group_bd_manager")
 
     def unlink(self):
         if self._dg_read_all_active():
             raise AccessError(_("Vous n'êtes pas autorisé à supprimer des opportunités/pistes CRM."))
+        if self._siha_agent_read_only():
+            raise AccessError(_("Le groupe SIHA AGENT n'est pas autorisé à supprimer des leads/opportunités."))
         if self._bd_manager_no_delete():
             raise AccessError(_("Le groupe BD MANAGER n'est pas autorisé à supprimer des leads/opportunités."))
         return super().unlink()
